Determine student affiliates via title or PU status
This allows people who have a title in STUDENT_TITLES to be
counted as students, even if their Princeton status is not
set to graduate or undergraduate.

For PGRAs, who should be listed on the Students page, this avoids
needing to set a "graduate" or "undergraduate" status, since
they are neither.

See #231