Add migration to prevent Project model deletion via cascade
Membership, Grant, and ProjectRelatedLink all have foreign keys
to Project and were set to be deleted via cascade when their
Project was deleted. This caused repeated exodus runs to clear out
these models, because the Projects would be deleted at the
beginning of the exodus for idempotency. This migration changes
the handler to SET_NULL; it can be changed back once v3 is released.

See also #285