Add and populate email for Person
Since moving away from the old proxy User model, we no longer
have email information stored on Person, which means it would
not be displayed in some cases on Profiles.

This commit adds an email field to Person and sets it to populate
based on the user's email when a user is initialized or updated
from LDAP. It also copies over emails during the migration from
User to Person so that no information is lost.