Make event speakers use through-model
The many-to-many relationship between Event and Person was causing
problems, likely because of modelcluster trying to aggregate people
at creation time, which threw 500 errors when making new Events.

This commit updates the relationship to use a new through-model
called Speaker, which is designed to mirror Author on BlogPost,
including the speaker_list convienience method and str() method.

Speakers are now editable using a better inline interface in the
admin and they can be used to search for an event in the admin
as well.