Use last_published_at timestamps as modification dates
For wagtail models, the built-in last_published_at timestamp
should be used as the most recent modification date for views
and templates that need this information. When a user updates a
Page model, we only care about when they *publish* changes to
that model, since those are the only modifications that show
to the public.

This commit adds migrations to remove the "updated" timestamp
fields from models that used them, and switch views to use the
"last_published_at" field instead. For Person, which was not a
Page, we keep the existing "updated_at" timestamp field.